Keep the debian links instead of recreating them
Recreating the link seems to have problems merging the sources
correctly. We ended up with a project.diff in the merged sources,
instead of project.diff being applied to generate the merged sources.

Anyway, copypac has an option to keep the project/package that link file
in the destination package points to, which does not seem to have the
same problem.